1、提取业务RMQ消息服务接口层 2、bug修复

+++ rmq-cms-api/src/main/java/com/cn/rmq/api/cms/model/vo/queue/CmsQueueVo.java
@@ -29,7 +29,7 @@ public class CmsQueueVo implements Serializable {
 
     private Integer checkDuration;
 
-    private Integer checkTimeout;
+    private Short checkTimeout;
 
     private LocalDateTime createTime;
